# tacocloud/tacocloud-api/src/main/java/tacos/api/dto/FavoriteResponse.java
package tacos.api.dto;

import java.time.Instant;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class FavoriteResponse {
    private String tacoId;
    private String tacoName;
    private Instant createdAt;
}