# tacocloud/tacocloud-api/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Taco Cloud API
  version: "1.0.0"
  description: |
    Contrato público v1 de Taco Cloud (TC-35).

    Versionado y deprecación: el contrato vigente vive en /api/v1. El alias legado
    /api sigue respondiendo igual durante la transición, pero cada respuesta del alias
    incluye los headers `Deprecation: true`, `Link: </api/v1/...>; rel="successor-version"`
    y `Warning: 299`. Los clientes deben migrar a /api/v1.

    Errores: todas las respuestas 4xx/5xx controladas usan `application/problem+json`
    con el schema ApiProblem. Política: 400 formato/validación, 401 sin autenticación,
    403 sin permiso, 404 no existe (o no es del usuario en rutas /users/me), 409 conflicto
    con el estado actual, 422 regla de negocio, 500 error inesperado sin detalles internos.

    Seguridad: HTTP Basic. Roles USER, ADMIN y KITCHEN (ver SecurityConfig).
servers:
  - url: /api/v1
    description: Contrato vigente
  - url: /api
    description: Alias legado (obsoleto)
security:
  - basicAuth: []
...
paths:
...
  /users/me/favorites:
    get:
      tags: [users-me]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Favoritos propios
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FavoritePage' }
  /users/me/favorites/{tacoId}:
    parameters:
      - { name: tacoId, in: path, required: true, schema: { type: string } }
    put:
      tags: [users-me]
      summary: Marcar favorito (idempotente)
      responses:
        '204': { description: Es favorito }
        '404': { $ref: '#/components/responses/Problem' }
    delete:
      tags: [users-me]
      summary: Quitar favorito (idempotente)
      responses:
        '204': { description: No es favorito }
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  parameters:
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
...
  responses:
    Problem:
      description: Error con estructura ApiProblem
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ApiProblem' }
  schemas:
    ApiProblem:
      type: object
      required: [type, title, status, instance, code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string }
        correlationId: { type: string }
        violations:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }
...
    FavoriteResponse:
      type: object
      properties:
        tacoId: { type: string }
        tacoName: { type: string }
        createdAt: { type: string, format: date-time }
    FavoritePage:
      type: object
      properties:
        content: { type: array, items: { $ref: '#/components/schemas/FavoriteResponse' } }
        page: { type: integer }
        size: { type: integer }
        hasNext: { type: boolean }
...